Added pop-up alert on deleting of a note

+++ views/note.view.php
@@ -5,7 +5,7 @@
 <p><?=$note['content']?></p>
 <p>Publiée le <?=$note['created_at']?> par <strong><?= $note['name']?></strong></p>
 
-<p><a href="/note-delete?id=<?=$note['id']?>">Supprimer cette note</a></p>
+<p><a href="/note-delete?id=<?=$note['id']?>" onClick="return confirm('Etes-vous certain de vouloir supprimer cette note?');">Supprimer cette note</a></p>
 
 
 <a href="/notes">Retour à la liste des notes</a>
